Improvements + bugfixes
Included the barrel length in the aimbot prediction, and added an automatic switch to punch chasing if target is too close. This aimbot prediction seems to be better then the other bots out there.

Also changed such that (shift <-> default) is the only key that needs to be pressed, hold that to lock aim onto a target and if the target is within punching distance, and if you have no weapons equipped, then it will start punch chasing. This is very useful against other bots.

Also fixed a bug in which if you activate the bot and someone is already in sight, then it might not identify you as the player correctly.